player: make sub-pos a float value
mpv makes this option an integer, but the underlying ass API actually
accepts doubles. From some testing, there is no meaningful precision
difference between float or double (it seems to go in roughly 0.05
steps), so just make it a float. sd_lavc also can handle non-integer
values here. Closes #11583.

+++ options/options.h
@@ -78,7 +78,7 @@ typedef struct mp_vo_opts {
 struct mp_subtitle_opts {
     bool sub_visibility;
     bool sec_sub_visibility;
-    int sub_pos;
+    float sub_pos;
     float sub_delay;
     float sub_fps;
     float sub_speed;
